# Improve performance creating TimeFromEpoch object