--- a/CHANGES.rst
+++ b/CHANGES.rst
@@ -55,6 +55,10 @@ astropy.tests
 astropy.time
 ^^^^^^^^^^^^
 
+- Improve initialization time by a factor of four when creating a scalar ``Time``
+  object in a format like ``unix`` or ``cxcsec`` (time delta from a reference
+  epoch time). [#10406]
+
 astropy.timeseries
 ^^^^^^^^^^^^^^^^^^