<!-- This comments are hidden when you submit the pull request,
so you do not need to remove them! -->

<!-- Please be sure to check out our contributing guidelines,
https://github.com/astropy/astropy/blob/master/CONTRIBUTING.md .
Please be sure to check out our code of conduct,
https://github.com/astropy/astropy/blob/master/CODE_OF_CONDUCT.md . -->

<!-- If you are new or need to be re-acquainted with Astropy
contributing workflow, please see
http://docs.astropy.org/en/latest/development/workflow/development_workflow.html .
There is even a practical example at
https://docs.astropy.org/en/latest/development/workflow/git_edit_workflow_examples.html#astropy-fix-example . -->

<!-- Astropy coding style guidelines can be found here:
https://docs.astropy.org/en/latest/development/codeguide.html#coding-style-conventions
Our testing infrastructure enforces to follow a subset of the PEP8 to be
followed. You can check locally whether your changes have followed these by
running the following command:

tox -e codestyle

-->

<!-- Please just have a quick search on GitHub to see if a similar
pull request has already been posted.
We have old closed pull requests that might provide useful code or ideas
that directly tie in with your pull request. -->

<!-- We have several automatic features that run when a pull request is open.
They can appear daunting but do not worry because maintainers will help
you navigate them, if necessary. -->

### Description
<!-- Provide a general description of what your pull request does.
Complete the following sentence and add relevant details as you see fit. -->

<!-- In addition please ensure that the pull request title is descriptive
and allows maintainers to infer the applicable subpackage(s). -->

This pull request is to address poor performance when creating a Time object in a format like `cxcsec` or `unix` using the `TimeFromEpoch` subclass.  The changes here were all driven by profiling with `prun` and `lprun`.

The biggest problem was unnecessarily creating two entirely separate `Time` objects, one for the `epoch` time with each new `Time` instance and a temporary one for doing the JD math. Instead this now does this once for the class and stores `epoch` as a common class attribute, and avoids the temporary one if the scale does not need changing (which is the most common case).

Other things:
- `day_frac` is slow (10 us) and can contribute noticeably to initialization time. I think there are opportunities to make this faster for common cases, e.g. `val2 == 0` or `0 <= val2 <= 1` (and `val1` a multiple of `0.5`.  Also cythonizing it seems like it might be worthwhile.
- `np.all()` is curiously slow for a scalar input.

#### Performance

See script below.

*Revision: 3ee79d1cd*
Scalar time = 64 us
Array time (1_000_000) = 91 us

*Revision: master*
Scalar time = 283 us
Array time (1_000_000) = 313 us

<!-- If the pull request closes any open issues you can add this.
If you replace <Issue Number> with a number, GitHub will automatically link it.
If this pull request is unrelated to any issues, please remove
the following line. -->

```
import time
import subprocess
import sys
import numpy as np

out = subprocess.run(['git', 'rev-parse', 'HEAD'], stdout=subprocess.PIPE)
curr_rev = out.stdout.decode('ascii')[:8]

rev = sys.argv[1]
subprocess.run(['git', 'checkout', rev])
print(f'Revision: {rev}')

from astropy.time import Time

n_loop = 10000
Time(1.5, format='cxcsec')
t0 = time.time()
for _ in range(n_loop):
    Time(1.5, format='cxcsec')
t1 = time.time()
print(f'Scalar time = {(t1 - t0) / n_loop * 1e6:.0f} us')

n_loop = 10000
n_times = 1_000_000
cxcsecs = np.arange(0.0, n_times, n_times, dtype=np.float64)
t0 = time.time()
for _ in range(n_loop):
    Time(cxcsecs, format='cxcsec')
t1 = time.time()
print(f'Array time (1_000_000) = {(t1 - t0) / n_loop * 1e6:.0f} us')

subprocess.run(['git', 'checkout', 'time-faster-epoch-time'])
```

Fixes #<Issue Number>