@pllim 2020-05-27
Is the failure related to the leap second stuff that @mhvk and @eteq were discussing? 🤷