@pllim 2020-05-27
Tests passed with minimal dependencies (Python 3.7) but failed with optional ones (Python 3.8). I recommend checking locally again with the same setup as the failed job.

Also the `html-docs` failure is just a PEP 8 thing. So you have two different things going on here.