@taldcroft 2020-05-28
@pllim - now to even *look at* the circleci: html-docs failure it wants me to authenticate, at which point it asks for invasive permissions for every single repo that I manage.  So how can I look at the HTML docs failure?