@pllim 2020-05-28
> So how can I look at the HTML docs failure?

I have mine hooked up to GitHub, but I am surprised that it asks for auth.