@pllim 2020-05-28
Anyway, here are the CircleCI errors:

**html-docs**
```
Warning, treated as error:
failed to import TimeFromEpoch.epoch
ERROR: InvocationError for command /home/circleci/project/.tox/build_docs/bin/sphinx-build -W -b html . _build/html (exited with code 2)
```

**32bit-and-parallel**
```
Error response from daemon: Get https://quay.io/v2/: EOF
```

The second one looks like server was down. Another push should resolve it. The doc failure seems related. Hope this helps!