@taldcroft 2020-05-28
Ugh. The `@classproperty(lazy_property=True)` is consistently giving this error when building sphinx docs:
```
WARNING: failed to import TimeFromEpoch.epoch
```
I tried removing the `lazy_property=True` and got the same result.  I have no clue how to resolve this sphinx error. Unless somebody else suggest something I will just go back to the first inelegant version of setting the property in `__init__`, but checking the class `__dict__` instead of using `getattr(cls, ...)` to avoid inheritence.