@pllim 2020-05-28
Is it coming from the API doc generation? If so, will excluding that property from API doc fix the problem? But then it won't be listed but you can maybe mention its existence in the class docstring or something. I can't tell where that warning is coming from exactly from the log.