@mhvk 2020-05-29
I'll have a look tomorrow. I'm pretty sure I use it in other packages with more success...