@taldcroft 2020-05-29
Yes, this is coming from the API docs and is related to the circular import of `Time` that happens at the end of `formats.py`. If I make the `epoch` class property just return `1` then the docs get generated without error. But trying something that actually passes tests still fails in doc generation:
```
    @classproperty
    def epoch(cls):
        if 'Time' in globals():
            return Time(cls.epoch_val, cls.epoch_val2, scale=cls.epoch_scale,
                        format=cls.epoch_format)
        else:
            return None
```

These days I would have documented why that `import Time, ..` statement is down at the bottom and how it actually manages to succeed, because I have forgotten both!