@pllim 2020-05-29
I tried several things but none of them worked. Each of these was tried separately:

* Use `warnings.simplefilter` to ignore all warnings on the import line at the end of `formats.py`.
* Use `autodoc_mock_imports` in `conf.py`.
* Use `suppress_warnings` in `conf.py`.
* Use `:skip:` in `time/index.py` where `automodapi` is used.

It could be that I wasn't using them correctly. Maybe @astrofrog has some tips on how to do this properly in `automodapi`.