@taldcroft 2020-05-29
@pllim - thanks for the valiant effort!! Much appreciated. Maybe getting rid of the circular imports (which I don't like anyway) and then my check on `'Time' in globals()` would work.  Hmm.