@taldcroft 2020-05-29
I have this passing local tests and sphinx generation now. I had to make a hidden class property and then a public instance property.  🤞on the official CI!