@bsipocz 2020-11-16
See the discussion about the backporting here: https://github.com/astropy/astropy/pull/10876#issuecomment-727551731